pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chemotaxis"
version = "0.1.0"
description = "A cell tracking program using ultrack and cellpose for automated live-cell imaging analysis"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "User", email = "user@example.com"}
]
keywords = ["cell-tracking", "microscopy", "cellpose", "ultrack"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "ultrack>=0.5.0",
    "cellpose>=3.0.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "scikit-image>=0.18.0",
    "tifffile>=2022.0.0",
    "pillow>=8.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
jupyter = [
    "jupyter>=1.0.0",
    "matplotlib>=3.5.0",
]

[project.scripts]
chemotaxis = "chemotaxis.main:app"

[tool.setuptools]
packages = ["chemotaxis"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/chemotaxis --cov-report=html --cov-report=term"
